[maxwell] Adding (fake) Modular Acquirer agent.
V2: Now with less breaking of the build!

This agent publishes the state of modular to the context service.
Currently, there is no way for SysUI or user_state_runner to
keep track of the status of all stories, so modular_state is stored as
0. However, in the future, we will be able to express the state of the
UI and replace that part of the modular_aqcuirer agent.

Change-Id: Iea35598ee05151f9ab9039729327d855e16fdf89