Adding test and fix for a feature omitted for 11/16
For 11/16, there was no need to clean up an Ask channel that had been
destroyed. Adding a test and handler for that.

Also revised the Next doc to reflect tested behavior, that killing the
controller should not kill the listener (a convenience feature for set-and-
forget controllers). Revised Ask doc to be consistent, and implemented with
that behavior in mind.

Also found a simplification that happens if added SuggestionPrototypes are
not const. Idk why I thought they should be const in the past.

Change-Id: I5ff9e54fc9a4469fd74798fbfb6917df3747ef45